refactor: home container folder renamed
