+++ src/pages/Home.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import HomeContainer from "components/home";
+import HomeContainer from "components/home-container";
 import Layout from "components/layout";
 
 function Home() {
